Working scala3 derivation code with single recursion for coproducts
