+++ core/src/main/scala-2/medeia/decoder/BsonDecoderVersionSpecific.scala
@@ -1,0 +1,7 @@
+package medeia.decoder
+
+import medeia.generic.GenericDecoder
+
+trait BsonDecoderVersionSpecific {
+  def derived[A](implicit genericDecoder: GenericDecoder[A]): BsonDecoder[A] = genericDecoder
+}
