+++ core/src/main/scala-2/medeia/encoder/BsonDocumentEncoderVersionSpecific.scala
@@ -1,0 +1,7 @@
+package medeia.encoder
+
+import medeia.generic.GenericEncoder
+
+trait BsonDocumentEncoderVersionSpecific {
+  def derived[A](implicit genericEncoder: GenericEncoder[A]): BsonDocumentEncoder[A] = genericEncoder
+}
