+++ core/src/main/scala-3/medeia/decoder/BsonDecoderVersionSpecific.scala
@@ -1,0 +1,7 @@
+package medeia.decoder
+
+import medeia.generic.GenericDecoder
+
+trait BsonDecoderVersionSpecific {
+    inline def derived[A]: BsonDecoder[A] = GenericDecoder[A]
+}
