+++ core/src/main/scala-3/medeia/encoder/BsonDocumentEncoderVersionSpecific.scala
@@ -1,0 +1,7 @@
+package medeia.encoder
+
+import medeia.generic.GenericEncoder
+
+trait BsonDocumentEncoderVersionSpecific {
+    inline def derived[A]: BsonDocumentEncoder[A] = GenericEncoder[A]
+}
