+++ core/src/main/scala-3/medeia/generic/GenericEncoderInstances.scala
@@ -1,13 +1,44 @@
 package medeia.generic
 
+import shapeless3.deriving.K0.*
 import shapeless3.deriving.*
+import org.mongodb.scala.bson.*
+import scala.compiletime.*
+import medeia.encoder.{BsonDocumentEncoder, BsonEncoder}
+import scala.deriving.Mirror
 
 trait GenericEncoderInstances {
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitConversion"))
-  implicit inline def genericEncoder[A](using
-      gen: K0.Generic[A],
-      gdOptions: GenericDerivationOptions[A] = GenericDerivationOptions[A](),
-      stOptions: SealedTraitDerivationOptions[A] = SealedTraitDerivationOptions[A]()
+  inline def apply[A]: GenericEncoder[A] =
+    summonInline[GenericEncoder[A]]
+
+  @SuppressWarnings(Array("org.wartremover.warts.SeqApply"))
+  given coproduct[A](using
+      inst: => K0.CoproductInstances[ProductEncoder, A],
+      labelling: Labelling[A],
+      mirror: Mirror.SumOf[A],
+      options: SealedTraitDerivationOptions[A] = SealedTraitDerivationOptions[A]()
   ): GenericEncoder[A] =
-    gen.derive(GenericProductEncoder.encoder, GenericCoproductEncoder.encoder)
+    (value: A) => {
+      inst.fold(value)(
+        [t] =>
+          (st: ProductEncoder[t], t: t) =>
+            val original: BsonDocument = st.encode(t).asDocument()
+            val label = labelling.elemLabels(mirror.ordinal(value))
+            original.append(options.discriminatorKey, BsonString(options.transformDiscriminator(label)))
+      )
+    }
+
+  given product[A: Labelling](using inst: => ProductInstances[BsonEncoder, A], options: GenericDerivationOptions[A] = GenericDerivationOptions[A]()): GenericEncoder[A] = ProductEncoder.product
+}
+
+private trait ProductEncoder[A] extends GenericEncoder[A]
+
+private object ProductEncoder {
+  given product[A](using inst: => ProductInstances[BsonEncoder, A], labelling: Labelling[A], options: GenericDerivationOptions[A] = GenericDerivationOptions[A]()): ProductEncoder[A] = (value: A) =>
+      BsonDocument(
+        labelling.elemLabels.zipWithIndex.map((label, i) =>
+          val fieldName = options.transformKeys(label)
+          fieldName -> inst.project(value)(i)([t] => (encoder: BsonEncoder[t], pt: t) => encoder.encode(pt))
+        )
+      )
 }
